session.goファイルに集約
